single UT for grpc
